% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_stabli.R
\docType{package}
\name{stabli-package}
\alias{stabli}
\alias{stabli-package}
\title{stabli: Statistical Tables and Lists Thereof}
\description{
\href{https://github.com/FinnishCancerRegistry/stabli/actions/workflows/R-CMD-check.yaml}{\if{html}{\figure{https://github.com/FinnishCancerRegistry/stabli/actions/workflows/R-CMD-check.yaml/badge.svg}}}
\subsection{Description}{

Create and make use of a more formal definition of a table of
statistics. Store metadata such as the names of stratifying and value
columns into the object itself so you don't have to keep track
yourself. Easily write functions to produce statistical tables.
}

\subsection{Installation}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{devtools::install_github("FinnishCancerRegistry/stabli@release")
}\if{html}{\out{</div>}}
}

\subsection{Example}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# stabli::stat_table_make_from_expr
my_stat_fun_1 <- function(
  x,
  by = NULL,
  subset = NULL,
  by_style = NULL
) \{
  out <- stabli:::stat_table_make_from_expr(
    expr = quote(list(n = .N, mu = mean(b))),
    meta_expr = quote(list(
      stratum_col_nms = as.character(names(by)),
      value_col_nms = c("n", "mu")
    )),
    dataset_nm = "x"
  )
  return(out[])
\}

my_stat_fun_2 <- function(
  x,
  by = NULL,
  subset = NULL,
  by_style = NULL
) \{
  out <- stabli:::stat_table_make_from_expr(
    expr = quote(\{
      dt <- data.table::data.table(
        interval_no = 1:5,
        time_lo = 0:4,
        time_up = 1:5,
        n = NA_integer_,
        d = NA_integer_
      )
      if (.N > 0) \{
        data.table::set(
          dt,
          j = c("n", "d"),
          value = list(
            n = sample(size = 5, x = 10L, replace = TRUE),
            d = sample(size = 5, x = 3L, replace = TRUE)
          )
        )
      \}
      dt[]
    \}),
    meta_expr = quote(list(
      stratum_col_nms = c(names(by), "interval_no"),
      value_col_nms = c("n", "d")
    )),
    dataset_nm = "x"
  )
  return(out[])
\}

# If you don't some args in your function
my_stat_fun_3 <- function(
  x,
  by = NULL
) \{
  subset <- NULL
  by_style <- NULL
  out <- stabli:::stat_table_make_from_expr(
    expr = quote(list(n = .N, mu = mean(b))),
    meta_expr = quote(list(
      stratum_col_nms = as.character(names(by)),
      value_col_nms = c("n", "mu")
    )),
    dataset_nm = "x"
  )
  return(out[])
\}

my_dataset <- data.table::data.table(
  a = sample(1:5, size = 1e6L, replace = TRUE)
)
my_dataset[j = "b" := a + runif(n = 1e6L)]

st_1 <- my_stat_fun_1(
  x = my_dataset,
  by = "a",
  subset = data.table::data.table(a = 3:5),
  by_style = "keep_empty"
)
stopifnot(nrow(st_1) == 5)

st_2 <- my_stat_fun_2(
  x = my_dataset,
  by = "a",
  subset = data.table::data.table(a = 3:5),
  by_style = "keep_empty"
)
stopifnot(nrow(st_2) == 5 * 5)

st_3 <- my_stat_fun_2(
  x = my_dataset,
  by = NULL,
  subset = data.table::data.table(a = 3:5),
  by_style = "keep_empty"
)
stopifnot(nrow(st_3) == 5)

st_4 <- my_stat_fun_1(
  x = my_dataset,
  by = NULL,
  subset = data.table::data.table(a = 3:5),
  by_style = "keep_empty"
)
stopifnot(nrow(st_4) == 1)

st_5 <- my_stat_fun_1(
  x = my_dataset,
  by = list(a = 1:6),
  subset = data.table::data.table(a = 3:5),
  by_style = "keep_empty"
)
stopifnot(nrow(st_5) == 6)

st_6 <- my_stat_fun_2(
  x = my_dataset,
  by = list(a = 1:6),
  subset = data.table::data.table(a = 3:5),
  by_style = "keep_empty"
)
stopifnot(nrow(st_6) == 6 * 5)

st_7 <- my_stat_fun_3(
  x = my_dataset,
  by = "a"
)
stopifnot(nrow(st_7) == 5)
}\if{html}{\out{</div>}}
}

\subsection{Releases}{

Run the numbered scripts in \verb{./dev/} to create a new release.
}
}
\section{News}{

\subsection{News for version 0.4.1}{
\subsection{stabli::print.stat_table_list}{

\code{stabli::print.stat_table_list} bug related to adding \code{stat_table}
metadata into the printed table fixed.
}

}

\subsection{News for version 0.4.0}{
\subsection{stabli::handle_arg_by}{

\code{stabli::handle_arg_by} gains argument \code{stratification_vame}.
}

\subsection{stabli::handle_arg_by_et_subset_et_by_style_inplace}{

\code{stabli::handle_arg_by_et_subset_et_by_style_inplace} gains argument
\code{handle_arg_nms}.
}

\subsection{stabli::handle_arg_by_et_subset_et_by_style_inplace}{

\code{stabli::handle_arg_by_et_subset_et_by_style_inplace} gains argument
\code{stratification_vame}.
}

\subsection{stabli::stat_table_list_make_from_settings}{

\code{stabli::stat_table_list_make_from_settings} arg \code{optional_steps} elem
\code{lapply_pre_stat_fun_call} renamed to \code{lapply_pre_stat_table_call}.
}

\subsection{stabli::stat_table_list_make_from_settings}{

\code{stabli::stat_table_list_make_from_settings} arg \code{optional_steps} elem
\code{lapply_post_stat_fun_call} renamed to \code{lapply_post_stat_table_call}.
}

\subsection{stabli::stat_table_make_from_by_list}{

\code{stabli::stat_table_make_from_by_list} arg \code{optional_steps} elem
gains element \code{lapply_pre_harmonisation_call}.
}

\subsection{stabli::stat_table_make_from_by_list}{

\code{stabli::stat_table_make_from_by_list} arg \code{optional_steps} elem
gains element \code{lapply_post_harmonisation_call}.
}

\subsection{stabli::stat_table_make_from_expr}{

\code{stabli::stat_table_make_from_expr} gains argument
\code{stratification_vame}.
}

}

\subsection{News for version 0.3.0}{
\subsection{stabli::assert_is_arg_by}{

New function \code{stabli::assert_is_arg_by}.
}

\subsection{stabli::assert_is_arg_by_style}{

New function \code{stabli::assert_is_arg_by_style}.
}

\subsection{stabli::assert_is_arg_subset}{

New function \code{stabli::assert_is_arg_subset}.
}

\subsection{stabli::handle_arg_by}{

New function \code{stabli::handle_arg_by}.
}

\subsection{stabli::handle_arg_by_et_subset_et_by_style_inplace}{

New function \code{stabli::handle_arg_by_et_subset_et_by_style_inplace}.
}

\subsection{stabli::handle_arg_by_style}{

New function \code{stabli::handle_arg_by_style}.
}

\subsection{stabli::handle_arg_subset}{

New function \code{stabli::handle_arg_subset}.
}

\subsection{stabli::stat_table_make_from_expr}{

New function \code{stabli::stat_table_make_from_expr}.
}

}

\subsection{News for version 0.2.0}{
\subsection{stabli}{

First release.
}

}

\subsection{News for version 0.1.0}{
\subsection{stabli::stat_table_list_make_from_settings}{

New exported fun \verb{[stabli::stat_table_list_make_from_settings]}.
}

}
}

\author{
\strong{Maintainer}: Joonas Miettinen \email{joonas.miettinen@cancer.fi} (\href{https://orcid.org/0000-0001-8624-6754}{ORCID})

}
\keyword{internal}
