% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_table_list_make.R
\name{stat_table_list_make}
\alias{stat_table_list_make}
\alias{stat_table_list_make_from_settings}
\title{\code{tab_list}}
\usage{
stat_table_list_make_from_settings(
  dataset,
  settings,
  optional_steps = NULL,
  harmonisation_vame = NULL,
  stratification_vame = NULL
)
}
\arguments{
\item{dataset}{\verb{[data.table]} (no default)

The main dataset.}

\item{settings}{\verb{[data.table]} (no default)

The settings table. We effectively pass each row of this table to
\code{stabli::stat_table_make_from_by_list}. Must have columns
\itemize{
\item \code{tab_nm} \verb{[character]}: Name of the table in the produced list.
E.g. \code{"count_time_series"}. Not passed to
\code{stabli::stat_table_make_from_by_list} but used to determine name of
output element.
\item \code{fun_nm} \verb{[character]}: Each \code{settings[["fun_nm"]][i]} is passed.
E.g. \code{"mypkg::myfun"}.
\item \code{by_list} \verb{[list]}: Each
\code{settings[["by_list"]][[i]]} passed.
\item \code{arg_list} \verb{[character, list]}: Each
\code{settings[["arg_list"]][i]} or \code{settings[["arg_list"]][[i]]} passed.
}}
}
\value{
\strong{stabli::stat_table_list_make_from_settings}

Returns a \code{stat_table_list}.
}
\description{
Functions related to \code{tab_list}.
}
\section{Functions}{

\strong{stabli::stat_table_list_make_from_settings}

Performs the following steps:
\itemize{
\item Calls \code{optional_steps[["on_entry"]](env = main_fun_env)} if that
\code{optional_steps} element exists. \code{main_fun_env} is the temporary evaluation
environment of \code{stabli::stat_table_list_make_from_settings}.
\item Calls \code{on.exit(optional_steps[["on_exit"]](env = main_fun_env))} if that
\code{optional_steps} element exists.
\item \code{lapply}'s through row numbers of \code{settings}. For each \code{i}:
\itemize{
\item Calls \code{optional_steps[["lapply_on_entry"]](env = lapply_fun_env)} if
that \code{optional_steps} element exists. \code{lapply_fun_env} is the evaluation
environment of the anonymous function used in the \code{lapply} call.
\item Calls \code{on.exit(optional_steps[["lapply_on_exit"]](env = lapply_fun_env))}
if that \code{optional_steps} element exists.
\item Collects \code{arg_list} for calling \code{stabli::stat_table_make_from_by_list}.
\code{settings[["arg_list"]][i]}
is parsed and evaluated if it is of class \code{character}. If it is a
quoted expression it is also evaluated first. The evaluation
environment is the environment where
\code{stabli::stat_table_list_make_from_settings} is called. The result of the evaluation
must be a \code{list} or \code{NULL}. Otherwise each
\code{settings[["arg_list"]]} element must be a \code{list} or \code{NULL}.
\item Calls \code{optional_steps[["lapply_pre_stat_table_call"]](env = lapply_fun_env)}
if that \code{optional_steps} element exists.
\item Calls \code{stabli::stat_table_make_from_by_list} with \code{arg_list_i}.
\item Calls \code{optional_steps[["lapply_post_stat_table_call"]](env = lapply_fun_env)}
if that \code{optional_steps} element exists.
\item Asserts that the produced object is of class \code{stat_table}.
See \verb{[stat_table]}.
}
\item Sets the names of the list produced by \code{lapply} to
\code{settings[["tab_nm"]]}.
\item Calls \code{stabli::stat_table_list_set} with
\code{meta = list(meta_dt = settings)} --- without column \code{tab_nm}.
}

Returns a \code{stat_table_list}.
}

\section{News for version 0.4.0}{


\code{stabli::stat_table_list_make_from_settings} arg \code{optional_steps} elem
\code{lapply_pre_stat_fun_call} renamed to \code{lapply_pre_stat_table_call}.

\code{stabli::stat_table_list_make_from_settings} arg \code{optional_steps} elem
\code{lapply_post_stat_fun_call} renamed to \code{lapply_post_stat_table_call}.
}

\section{News for version 0.1.0}{


New exported fun \verb{[stabli::stat_table_list_make_from_settings]}.
}

