% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_table_list_meta.R
\name{stat_table_list_meta}
\alias{stat_table_list_meta}
\alias{stat_table_list_meta_name}
\alias{stat_table_list_meta_default}
\alias{stat_table_list_meta_assert}
\alias{stat_table_list_meta_get}
\alias{stat_table_list_meta_set}
\alias{stat_table_list_class_name}
\alias{stat_table_list_class_set}
\title{\code{stat_table_list} Metadata}
\usage{
stat_table_list_meta_name()

stat_table_list_meta_default()

stat_table_list_meta_assert(
  x,
  x_nm = NULL,
  call = NULL,
  assertion_type = NULL,
  stat_table_list = NULL
)

stat_table_list_meta_get(x)

stat_table_list_meta_set(x, meta)

stat_table_list_class_name()

stat_table_list_class_set(x)
}
\arguments{
\item{x}{\verb{[stat_table_list]} (no default)

A \code{stat_table_list} object.}

\item{x_nm}{See \verb{[dbc::handle_args_inplace]}.}

\item{call}{See \verb{[dbc::handle_args_inplace]}.}

\item{assertion_type}{See \verb{[dbc::handle_args_inplace]}.}

\item{stat_table_list}{\verb{[NULL, integer]} (default \code{NULL})

Check \code{nrow(x[["meta_dt"]])} against this value if given where
\code{x} is the metadata object of a \code{stat_table_list}.
\itemize{
\item \code{NULL}: No check is performed.
\item \code{integer}: This is checked against.
}}

\item{meta}{\verb{[list]} (no default)

Metadata object for a \code{stat_table_list}. See what
\code{stabli::stat_table_list_meta_assert} requires it to look like.}
}
\description{
Handle the metadata of a \code{stat_table_list}.
}
\section{Functions}{

\strong{stabli::stat_table_list_meta_name}

Get name of \code{stat_table_list} attribute containing its metadata.


\strong{stabli::stat_table_list_meta_default}

Get default metadata object for a \code{stat_table_list} object.


\strong{stabli::stat_table_list_meta_assert}

The metadata object of a \code{stat_table_list} must pass these checks:
\itemize{
\item Must be a named \code{list} object.
\item If the metadata object contains element \code{meta_dt}, it must be a
\code{data.table} object with as many rows as the \code{stat_table_list} object
has elements.
}


\strong{stabli::stat_table_list_meta_get}

Get \code{stat_table_list} metadata.


\strong{stabli::stat_table_list_meta_set}

Set (assign) \code{stat_table_list} metadata.


\strong{stabli::stat_table_list_class_name}

Get name of \code{stat_table_list} class. Returns
\code{"stat_table_list"}.


\strong{stabli::stat_table_list_class_set}

Set class of a \code{stat_table_list} object.
}

