% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_stabli.R
\docType{package}
\name{stabli-package}
\alias{stabli}
\alias{stabli-package}
\title{stabli: Statistical Tables and Lists Thereof}
\description{
\href{https://github.com/FinnishCancerRegistry/stabli/actions/workflows/R-CMD-check.yaml}{\if{html}{\figure{https://github.com/FinnishCancerRegistry/stabli/actions/workflows/R-CMD-check.yaml/badge.svg}}}
\subsection{Description}{

Create and make use of a more formal definition of a table of
statistics. Store metadata such as the names of stratifying and value
columns into the object itself so you don't have to keep track
yourself.
}

\subsection{Example}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{st1 <- stabli::stat_table(
  data.table::data.table(a = 1:3, v = 3:1),
  list(
    stratum_col_nms = "a",
    value_col_nms = "v"
  )
)
st2 <- stabli::stat_table(
  data.table::data.table(b = 1:3, v = 3:1),
  list(
    stratum_col_nms = "b",
    value_col_nms = "v"
  )
)
stl <- stabli::stat_table_list(
  list(tab1 = st1, tab2 = st2),
  list(meta_dt = data.table::data.table(stat_fun_nm = c("fun1", "fun1")))
)
}\if{html}{\out{</div>}}
}

\subsection{Releases}{

Run the numbered scripts in \verb{./dev/} to create a new release.
}
}
\section{News}{

\subsection{News for version 0.3.0}{
\subsection{stabli::assert_is_arg_by}{

New function \code{stabli::assert_is_arg_by}.
}

\subsection{stabli::assert_is_arg_by_style}{

New function \code{stabli::assert_is_arg_by_style}.
}

\subsection{stabli::assert_is_arg_subset}{

New function \code{stabli::assert_is_arg_subset}.
}

\subsection{stabli::handle_arg_by}{

New function \code{stabli::handle_arg_by}.
}

\subsection{stabli::handle_arg_by_et_subset_et_by_style_inplace}{

New function \code{stabli::handle_arg_by_et_subset_et_by_style_inplace}.
}

\subsection{stabli::handle_arg_by_style}{

New function \code{stabli::handle_arg_by_style}.
}

\subsection{stabli::handle_arg_subset}{

New function \code{stabli::handle_arg_subset}.
}

\subsection{stabli::stat_table_make_from_expr}{

New function \code{stabli::stat_table_make_from_expr}.
}

}

\subsection{News for version 0.2.0}{
\subsection{stabli}{

First release.
}

}
}

\author{
\strong{Maintainer}: Joonas Miettinen \email{joonas.miettinen@cancer.fi} (\href{https://orcid.org/0000-0001-8624-6754}{ORCID})

}
\keyword{internal}
